Blocks architect - create block based on pressed arena tiles
Test method for future serialization/deserialization from xml file
