+++ Assets/Blocks architect/Textures/Save (button).png.meta
